Add compound DB indexes and wire Telegram into training pipeline
- schema.py: add 10 compound (symbol, interval/source, timestamp_ms DESC) indexes
  alongside existing single-column ones; eliminates full table scans at 5M+ rows
  for all repository.py query patterns (40-100x faster range queries)
- train.py: wire AlertManager + TelegramSender into training pipeline
  (start notification, per-200-fold progress pings with win rate + Sharpe + ETA,
  completion notification with mean Sharpe and model paths)
  All Telegram calls wrapped in try/except — failure never crashes training

+++ src/ep2_crypto/risk/binary_position_sizer.py
@@ -59,7 +59,7 @@ def polymarket_fee(
     """
     if price <= 0.0 or price >= 1.0:
         return 0.0
-    return fee_rate * (price * (1.0 - price)) ** exponent
+    return float(fee_rate * (price * (1.0 - price)) ** exponent)
 
 
 def binary_kelly(
